ml-service/requirements.txt
flask==3.0.3
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
pillow==10.2.0
thefuzz==0.22.1
python-Levenshtein==0.23.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2







